Ensure return types match fn's
Functions were permitted to be annotated with one return type, but return another. This should not be allowed.

It might be desirable in the future to allow return of None from all functions, giving a sort of Optional type behavior, but that would require changing the return type to something like a pointer which can communicate the lack of a value.